bun all-in-one
repale esbuild and http-zerver with BUN only. ony two tools will be Zig and Bun.

in future would love to bring back the http-zerver. I will still use it in smaller proejcts but would like to keep dependancies small here so if we already have bun may as well use it